renamed route to receive_route
